pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twindb-lightrag-memgraph"
version = "0.2.5"
description = "Memgraph KV, Vector, and DocStatus storage backends for LightRAG"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "lightrag-hku>=1.4.9,<2.0.0",
    "neo4j>=5.0.0,<7.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
